add update rest endpoint for editing polls

+++ app/Http/Controllers/PollController.php
@@ -39,38 +39,75 @@ class PollController extends Controller
             'is_anonymous' => $data['isAnonymous'],
             'user_id' => $userId
         ]);
-        $options = array_map(function ($text, $idx) {
-            return new Option([
-                'index' => $idx,
-                'text' => $text,
-            ]);
-        },
-            $data['options'],
-            array_keys($data['options'])
-        );
-        DB::transaction(function () use ($poll, $options) {
+        DB::transaction(function () use ($poll, $data) {
             $poll->save();
-            array_map(function ($option) use ($poll) {
-                $option->poll_id = $poll->id;
-                $option->save();
+            $options = array_map(function ($text, $idx) use ($poll) {
+                return [
+                    'index' => $idx,
+                    'text' => $text,
+                    'poll_id' => $poll->id
+                ];
             },
-                $options
+                $data['options'],
+                array_keys($data['options'])
             );
+            Option::insert($options);
         });
         return response()->json(['success' => 'success'], 200);
     }
 
-    public function update(Request $request)
+    public function update(Request $request, $id)
     {
+        $user_id = Auth::user()->id;
+        $poll = Poll::find($id);
+        if ($poll === null || $poll->user_id != $user_id) {
+            return response()->json(['error' => 'invalid poll id'], 403);
+        }
+        $data = $request->validate([
+            'title' => 'required|string|filled',
+            'options' => 'required|array',
+        ]);
+        $indices = array_map(function ($option) {
+            return $option['index'];
+        }, $data['options']);
 
+        DB::transaction(function () use ($data, $indices, $id, $poll) {
+            Option::where([
+                ['poll_id', '=', $id],
+            ])->whereNotIn('index', $indices)->delete();
+
+            array_map(function ($option) use ($id) {
+                $result = Option::where([
+                    ['poll_id', '=', $id],
+                    ['index', '=', $option['index']]
+                ])->first();
+                if ($result === null) {
+                    $opt = new Option([
+                        'poll_id' => $id,
+                        'index' => $option['index'],
+                        'text' => $option['text'],
+                        'vote_count' => 0
+                    ]);
+                    $opt->save();
+                } else {
+                    DB::update(
+                        "UPDATE options SET text = ? WHERE poll_id = ? AND `index` = ?",
+                        [$option['text'], $id, $option['index']]
+                    );
+                }
+            }, $data['options']);
+            $poll->title = $data['title'];
+            $poll->save();
+        });
+        return response()->json(['success' => 'success']);
     }
 
     public function destroy($id)
     {
         $user_id = Auth::user()->id;
         $poll = Poll::find($id);
-        if ($poll->user_id != $user_id) {
-            return response()->json(['error' => 'user did not create this poll'], 403);
+        if ($poll === null || $poll->user_id != $user_id) {
+            return response()->json(['error' => 'invalid poll id'], 403);
         }
         $poll->delete();
         return response()->json(['poll_id' => $poll->id]);
